# front_end/front_end/swagger/swagger_timeslot_service.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "cs9322 assign2 timeslot_service",
    "contact": {
      "email": "zack@nono.fi"
    }
  },
  "host": "timeslot_service",
  "basePath": "/",
  "tags": [
    {
      "name": "timeslot",
      "description": "check business time and make appointment"
    },
    {
      "name": "booking",
      "description": "check appointment status and cancel"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/timeslot": {
      "get": {
        "tags": [
          "timeslot"
        ],
        "summary": "ask for business time",
        "operationId": "askbusinesstime",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Usually, our dentists are available for one-hour appointment from 9:00 AM to 5:00 PM every day"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "timeslot"
        ],
        "summary": "make an appointment",
        "operationId": "makeanappointment",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "2dca069a90bcfd19367047bc96501ee1d23a72ce4a89b67d5376ae88f9e45fed"
                },
                "dentist_id": {
                  "type": "integer",
                  "example": "5"
                },
                "ts": {
                  "type": "string",
                  "example": "2020-11-19T12:00:00.000+11:00"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "success"
                }
              }
            }
          },
          "409": {
            "description": "The requested time slot is booked by the other user. Conflict",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "This timeslot has been taken."
                },
                "available_timeslot": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "2020-11-19T11:00:00.000+11:00",
                    "2020-11-19T13:00:00.000+11:00",
                    "2020-11-19T14:00:00.000+11:00"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/booking": {
      "get": {
        "tags": [
          "booking"
        ],
        "summary": "check booking status",
        "operationId": "checkbookingstatus",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "description": "session 2dca069a90bcfd19367047bc96501ee1d23a72ce4a89b67d5376ae88f9e45fed"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "You have an appointment with Dr Lee 18/11/2020 4PM (CANCELLED) You have an appointment with Dr Lee 19/11/2020 10AM"
                }
              }
            }
          },
          "404": {
            "description": "appointment not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "You don't have any available appointment."
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "booking"
        ],
        "summary": "cancel appointment",
        "operationId": "cancelappointment",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "description": "session 2dca069a90bcfd19367047bc96501ee1d23a72ce4a89b67d5376ae88f9e45fed"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Your appointment has been cancelled. Say 'check appointment' if need to check appointment status."
                }
              }
            }
          },
          "404": {
            "description": "appointment not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "You don't have any available appointment."
                }
              }
            }
          }
        }
      }
    }
  }
}